refactor(pgvector): extract PostgreSQLDocumentStore base class
Move all SQL constants and data-layer methods into a new
`_base.py` abstract base class (`PostgreSQLDocumentStore`).
`PgvectorDocumentStore` is now a thin subclass that only
implements `_ensure_db_setup` and `_ensure_db_setup_async`
using psycopg.

Future PostgreSQL-backed integrations (AlloyDB, etc.) can
subclass `PostgreSQLDocumentStore` and override only the
connection layer, inheriting all SQL/data logic for free.

Closes #3239
