+++ integrations/pgvector/src/haystack_integrations/document_stores/pgvector/__init__.py
@@ -1,6 +1,7 @@
 # SPDX-FileCopyrightText: 2023-present deepset GmbH <info@deepset.ai>
 #
 # SPDX-License-Identifier: Apache-2.0
+from ._base import PostgreSQLDocumentStore
 from .document_store import PgvectorDocumentStore
 
-__all__ = ["PgvectorDocumentStore"]
+__all__ = ["PostgreSQLDocumentStore", "PgvectorDocumentStore"]
